data_pipeline/dbt-dag/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos
apache-airflow-providers-snowflake
proto-plus==1.22.3
protobuf==4.25.3
google-cloud-secret-manager==2.16.3